fix finalize contract checker (#309)

+++ user-service/src/main/java/rs/edu/raf/banka1/services/implementations/ContractServiceImpl.java
@@ -133,6 +133,8 @@ public class ContractServiceImpl implements ContractService {
         Contract contract = contractRepository.findById(contractId).orElseThrow(() -> new ContractNotFoundByIdException(contractId));
         if(!contract.getBankApproval() || !contract.getSellerApproval()) return;
 
+        if(contract.getRealizationDate() != null) return;
+
         //Transfer capital
         capitalService.removeFromPublicCapital(contract.getListingId(), contract.getListingType(), contract.getSeller(), contract.getAmount());
         capitalService.removeBalance(contract.getListingId(), contract.getListingType(), contract.getSeller(), contract.getAmount());
@@ -141,6 +143,8 @@ public class ContractServiceImpl implements ContractService {
         //Transfer funds
         bankAccountService.removeBalance(contract.getBuyer(), contract.getPrice());
         bankAccountService.addBalance(contract.getSeller(), contract.getPrice());
+
+        contract.setRealizationDate(Instant.now().toEpochMilli());
     }
 
 }
